server/requirements.txt
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
channels==1.0.3
constantly==15.1.0
coverage==4.3.4
daphne==1.0.2
Django==1.10
django-cors-headers==2.0.2
django-friendship==1.5.0
django-nose==1.4.4
django-redis==4.7.0
djangorestframework==3.5.3
incremental==16.10.1
msgpack-python==0.4.8
nose==1.3.7
packaging==16.8
pyparsing==2.1.10
redis==2.10.5
six==1.10.0
Twisted==16.6.0
txaio==2.6.0
zope.interface==4.3.3
